+++ Tests/SwiftLayoutTests/SwiftLayoutTests.swift
@@ -353,52 +353,3 @@ extension UIAccessibilityIdentification {
         Tag(taggable: self)
     }
 }
-
-
-protocol _Layout {
-    func active()
-    func attachSuperview(_ superview: UIView?)
-    func attachSuperview()
-}
-
-extension _Layout {
-    func active() {
-        attachSuperview()
-    }
-    func attachSuperview() {
-        attachSuperview(nil)
-    }
-}
-
-protocol _LayoutContainable: _Layout {
-    var layouts: [_Layout] { get }
-}
-
-extension _LayoutContainable {
-    func attachSuperview(_ superview: UIView?) {
-        for layout in layouts {
-            layout.attachSuperview(superview)
-        }
-    }
-}
-
-protocol _LayoutViewContainable: _LayoutContainable {
-    var view: UIView { get }
-}
-
-extension _LayoutViewContainable {
-    func attachSuperview(_ superview: UIView?) {
-        superview?.addSubview(view)
-    }
-}
-
-struct LayoutViewContainer: _LayoutViewContainable {
-    var view: UIView
-    var layouts: [_Layout]
-}
-
-extension Array where Element: _LayoutContainable {
-    var layouts: [_Layout] {
-        self
-    }
-}
